Introduce successorsOf(), predecessorsOf() on Digraph.
This also meant finally fixing the incorrect implementation of adjacentTo on digraphs.
